Rename logging module to avoid built-in conflict
